day_4/book-app/requirements.txt
Django
djangorestframework
drf-yasg